Remove unneccessary test
This test looks at the scenario when the AWS console shows the
email as not verified, but in our database the email is verified.

This can never happen as we cannot change the verification status
to be unverified, ie: change the sending_enabled status.
We only have the power to delete or create email identities.
Therefore, we can remove this test.