Add email templates and ConfirmationEmailSettings class
The email template will need to be able to iterate through
the email component ids to allow the form editor to pick
which email component will be the one to use as the confirmation
email.

ConfirmationEmailSettings class will allow us to get the correct
values that are required for the confirmation email feature. These
values are primarily gathered via a form and can be adjusted per
deployment environment.