Add delete modal for email questions
Similar to blocking on a page level, we want to prevent users from
deleting the specific email component question on a mulitple
question page.

Since this modal triggers on a component level rather than a page
level, users will be able to delete other components on their
mulitple question page.