Update modal focus() to allow for links or buttons.
The confirmation email question deletion modal has its confirmation button as a link to the settings page,
so this change allows that button to correctly gain focus on opening the modal.

Also includes a small change to the wording in the modal, and a change to the worduing in a similar branching
modal for consistency.