Upload service metadata to S3 during publishing
We have finally reached the point where the size of the metadata for
forms being injected into containers as an environment variable has
become too large.

To get round this issue, during the pre_publishing step we upload both
the service metadata and any autocomplete items associated with
autocompete components to an S3 bucket in the given platform namespace:

For test environments these are:

- formbuilder-platform-test-dev
- formbuilder-platform-test-production

For live environments these are:

- formbuilder-platform-live-dev
- formbuilder-platform-live-production

Yes these namespace names are confusing.

Currently the credentials for each of these buckects are set manually as
a secret in Kubernetes however they will in the near future be moved
into to match the way other secrets are injected into the Editor.

We have also removed the injection of the service metadata and
autocomplete items into the environment of the form containers entirely.
Each form will make a request to the S3 bucket to retrieve two objects:

- <service_id>_metadata.json
- <service_id>_autocomplete_items.json