Ensure confirmation email modal is prioritised over branch modals
Currently, when a user tries to delete a page that has an email
component used in confirmation email and is a destination of a
branch, the user is presented with the delete destination page
modal. This modal allows the user to delete the page, this is not
what we want, instead we want the confirmation email modal to
appear as this blocks the user from deleting the page until they
have changed their confirmation email settings.
This commit implements this change.