Set up the ReferenceNumberUpdater and ServiceConfiguration
We will inject the Runner with the 'REFERENCE_NUMBER' config
if the user enables this setting.
If the setting is disabled, 'REFERENCE_NUMBER' will not be injected
into the Runner.

We want to update this config globally for both dev and production
environments as the email templates could be different for each
environment.

Co-authored-by: Hellema Ibrahim <hellema.ibrahim@digital.justice.gov.uk>