Use the correct default depending on if reference number is enabled
Use to default values in email templates when ref num disabled.
Use reference number default email templates to the locales
and use these when reference number is enabled.

Refactor the updater to pass in the correct locale as an argument
when creating or updating the service configs.

Update locales for reference number placeholder