Interpolate reference number with reference number sentence
With the implementation of reference number, we need to be able to
provide different service email and confirmation email defaults
depending on whether reference number is enabled or disabled.

To achieve this, we have changed the default values by inserting
`{{reference_number_sentence}}` and gsubbing this with either a
sentence or an empty string.

This broke many tests, that have also been fixed in this commit.

Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>
Co-authored-by: Hellema Ibrahim <hellema.ibrahim@digital.justice.gov.uk>