Update first_publish? method
Passing only published records into the presenter would mean that we returned true
when it shouldn't.  This was done to try and show the first_pulbish method after the form
had been unpublished.

To achieve this a check for unpublished records has been added, and if there are any, we check for published
records since the last unpublish.  IF there are none, then its another "first" publish.