Refactor message display
If the form was published for the first time and has only been published once, `first_publish?` would
always be true, meaning that the default 15min warning would show on every visit to the publishing page.

To solve this, we do an initial get request in JS - if that succeeds the service is already live, so we
don't show the message. If it fails, we show the default message and start polling the url.  Once we
receive a 200 response, we show the success message.

Because we wait on a HTTP request, we can reduce the initial setTimeout delay.

Now that the JS does the first publish check, we no longer need the (slightly awkward)
first_publish? method in the presenter, as we can always have the partial present in
the view.