Remove placeholder for default values
When a user first creates a form, the editor uses the default
values, which now have the placeholder.

Add a method `assign` in the ContentSubstitutor that will return
the correct content based on whether the reference_number is enabled.

Also create a ContentSubstitutorHelper module as the ContentSubstitutor
is now being used in a number of places.