Reinstate first_publish? method in controller
It seems that relying solely on the JS response is not relaible - chrome specifically always seems to return a 404 on page load
even for a published service - causing the message to show. For improved reliability switch back to a backend first_publish?
method, which means the first publish message partuial (and js) is only in the page on first publish of the form.