Show autocomplete warning for pages in main flow only
For any pages that have an autocomplete page we show a warning on the
Publishing page if the user has not uploaded an autocomplete items for
any of the autocomplete components in their form.

The Editor also allows pages to become unconnected (detached) should the user
decide to do so. The only pages that are followed by the published form
are those that are in the main flow therefore there should be no way for
a public citizen to reach the unconnected page, unless they magically
know its' URL.

Therefore there is no need to show a publishing warning for any
autocomplete pages without autocomplete items that do NOT exist in the
main flow.

To achieve this, pass the MetadataPresenter::Grid object into the
AutocompleteItemsPresenter as the Grid object knows which pages are in
the main flow.