Set a href tag in the confirmation email body
If the user has enabled confirmation emails there will be place holder
text {{payment_link}} in the body. We want to replace this with the full
a tag element so that the link can be clickable once it has been
processed by the Submitter.

Both the {{payment_link}} placeholders get gsub'd and replaced with the
payment link concatenated with the reference number in the runner before
being sent onto the submitter.