Remove and recreate the email identity on AWS when it becomes 'unverfied'
This is to address a bug whereby a user adds a from address email,
the email is not verified within the 24hr timeframe, the identity
status on AWS changes to 'unverified'. When the user attempts to
save the same email address again, it does not fall into any of
the conditions in the `verify_email` method, since the email is:
- Not new so `get_email_identity` will return somethings
- The email was never verified, so `sending_enabled` is not true

Since it does not fall into either condition, the status of will
become `nil` in our database, which in turn will not trigger the
notification banner in the templates. So the user cannot resend
the validation link.

This commit fixes this by adding an extra condition, checking for
if the email identity is present but it has not yet been verified.
In this case we ensure the status is `pending`.