Ensure payment url link is stripped of whitespace
We want to silently remove the whitespace in the a payment link
url. Previously, we were only removing it on the client side but
still saving the whitespace in our databse.
This commit, fixes this issue and ensures we strip the whitespace
in the controller before saving it to the database.