Move allowed domains into a custom config
We have two domain allow lists that hold the same email domains.
Since we are about to add a third domain level allow list, that
will also hold the same domains, we should move the list into a
helper. This means we will have a single domain allow list and
reduce repetition.

This commit moves the domains to a custom config and changes the
auth0 and from_address class to reference the new config.