Domain validator for send to
We want to ensure the email address provided for the send to
feature has a domain that is on our allow list.

If it is on the allow list then we can save it to the
ServiceConfiguration table. If it is not on the allow list, we
want to not save it and show an error message.