Update 'collect information by email' settings page
As part of reply-to, we will now use the default no-reply-moj-forms
email address as the 'From address'.
So we no longer want users to access the 'From Address' page via the
'Change from address' link. This commit removes access to the link
and we now just present the Form name, not the email address.