Fix acceptance tests
The confirmation email settings now requires the user to add an
email address.

Previously, the acceptance tests assumed clicking save would mean
confirmation emails had been configured. But in these acceptance
tests it just triggers a validation error and hence the messaging
on the publishing page and reference payment setting pages were
not as expected. Updated the acceptance tests so we are now
submitting an email for the confirmation email settings page.