Remove FromAddress model from test services configs remover
The Test Services Configs Remover is a rake task that removes all the
configurations for a services that are not owned by the MoJ Forms team.
For the most part, this removes configs created by acceptance tests and
former members of the MoJ Forms team.

This rake task runs everyday at 5am, and hence can be removed as we
released Reply-To last week, so there shouldn't be any from address
configs anyway.