Add a deprecation note on the AWS SES Client
Now that we do not validate from address emails, we no longer call
out to AWS SES and hence, no longer use this adapter.

We have decided to add a deprecation message as we know we will
need AWS SES in the near future.

We are deleting the dummy adapter as it is no longer needed.