Add save and return service config to forms deployed to 'dev'
Injecting save and return config means we can leverage the current
submission mechanism and as a result won't need to make too many
changes down the road in the Runner / Submitter.

Since the template requires the service name, we need to pass this
into the PublishServiceCreation class, which is the point just
before the point of service publishing.