Increase the read timeout but keep the open timeout as is
We have an error which only affects the read timeout. This has
occurred when a user has tried to publish a form to the Test env.

We think because the form trying to be published is very large,
it is taking too long to read the metadata from the metadata api
and then timing out.

Only changing the read timeout as the open timeout is not the issue.

Faraday's `open_timeout` is equivalent to Net::HTTP's `open_timeout`.
Faraday's `timeout` is equivalent to Net::HTTP's `read_timeout`.

Open timeout: Number of seconds to wait for the connection to open.
Read timeout: Number of seconds to wait for one block to be read.