Fix Editor preview footer link bug
Currently, when we are on a page other than the start page in
preview mode and we click on any of the footer links, we receive
an error.

This is because the presenter gem would check if the editor url
ended in 'preview' to ascertain whether the editor was in preview
and then construct the footer links accordingly.

Bump the presenter gem to fix this error.