Add Save and Return Settings page
Users should be able to enable or disable the Save and Return feature.

This commit creates the settings page and the ability to update the
`SAVE_AND_RETURN` service config.

If the save and return config is enabled, a `SAVE_AND_RETURN`
and 'SAVE_AND_RETURN_EMAIL' ServiceConfiguration will be created.
If the save and return config is disabled, these configs will be
deleted.