Create PREVIOUS_SERVICE_SLUG if form has been published
We don't want to overwrite the existing service slug config if
the form has been published as we will need this to unpublish any
test forms.

We want to create a PREVIOUS_SERVICE_SLUG if:
- a form has been published
- there is no exisiting PREVIOUS_SERVICE_SLUG record.

This is because we do not want to overwrite the PREVIOUS_SERVICE_SLUG
config for every update.