Create a SERVICE_SLUG if one does not exist
Forms that were created when form name and url were coupled will
not have a SERVICE_SLUG config.

To fix this, we will create a SERVICE_SLUG on the settings page
if one does not exist. This is because we require a SERVICE_SLUG
config to create a PREVIOUS_SERVICE_SLUG if the form has already
been published.