Pass grid object into presenters
On the publishing page, we run a number of checks to ensure the
form can be pubished, eg: if there is a CYA/confirmation page in
the main flow.

These checks are done through presenters and each presenter creates
its own grid object. This meant on larger forms, the publishing page
would timeout due to having to build the grid object 2-3 times.

To fix this, we use the same grid object and pass this around to
each presenter. This means the grid object is only created once
when visiting the publishing page.

Co-authored-by: Chris Pymm <chris.pymm@digital.justice.gov.uk>
Co-authored-by: Hellema Ibrahim <hellema.ibrahim@digital.justice.gov.uk>
Co-authored-by: Matt Tei <matt.tei@digital.justice.gov.uk>