ComponentExpression model
This model ensures we return the correct expressions and expression
answers, when a question is selected on the conditional content
modal.

For now, we only support radios and checkboxes in conditional
content, and these are the only questions that should return
expressions and expression answers.

Since this is similar to branching, we have re-used the
`supports_branching?` method, to check the component type is either
radio or checkbox.

We also have different operators depending on whether it is radio
or checkbox - this is addressed in the `operators` method.

We also do not return any expression answers if the chosen operator
is `answered` or `not_answered`.

This class is largely similar to the `Expression` model used for
branching - we can look at refactoring once we have the feature
in place.