ComponentConditional model
The ComponentConditional is an object used to encapsulate a number
of ComponentExpressions.

A ComponentConditional uses an 'if' or and 'and' conditional type.
If there is more than one ComponentExpression in the
ComponentConditional, we would use the 'and' type.

Each ComponentConditional has a uuid, this is to allow us to
modify/update and delete the ComponentConditional.