wip: comment out removers
This assumes we have an li item that has the class:
"govuk-link destructive branch-remover"

It removes this line item (which provides the button to delete the
rule) if there is only one rule on the modal.

Since we are not worrying about deleting rules at the moment, I
am commenting this out for now.