wip: add edit template and create/post route
The post route was drawn without the :component_uuid in the param,
so I had to explicitly write it out.

Created the ContentVisibilityCreation as a way to create the new
metadata version with the conditionals - this is not yet working
as the expressions for some reason are not sending the operators.