Add required dummy env variables
As a side effect of the previous PR #2535, the assets precompile rake task that runs as part of the docker image build process requires the encryption env variables to be present, however just dummies.

All real env variables are injected into the kubernetes pod environment so shouldn't have any other effect to the best of my understanding.