Simple feature flag mechanism per environment
This is a simple feature flag mechanism I've used in other services, with little modifications.

This will allow us to check if some code should run or not depending if we are on local environments (including CI and rspec, etc.) or on kubernetes `formbuilder-saas-test` or `formbuilder-saas-live`.

Its main purpose is to handle more easily the address component visibility.

Co-Authored-By: H3ll3m4 <26165112+H3ll3m4@users.noreply.github.com>