Fix blue notification appearing on form error
https://trello.com/c/kjGWZoFD

The blue important message box was showing when there is an error message even though it's not supposed to.

This is because on form submission some variables were not being properly assigned and if the form gave error, the view would incorrectly show the blue notification due to this.

Fixed by assigning these variables also in the `#create` action so they are present in the view if the form triggers an error.