Announcements admin scaffolding
https://trello.com/c/guzlAuBY

First part of the notifications banner work.

This PR prepares the admin to perform basic CRUD operations on the announcement records.

I've also cleaned up old custom administrate views in favour of the bundled ones as there was no changes. Where there are changes, like for users or now announcements, namespaces views are created that don't interfere with the other resources.