Implement announcement dismiss functionality
https://trello.com/c/BX3rrOMO

Follow-up PR to continue with the notification banners work.

Implement the dismiss functionality so that users who don't want the notification showing all the time can hide it and will not show up again (until a new announcement is pushed in the future).

This is database-backed, so dismissals performed in a browser, will be honoured in a different browser, and the notification will not show.