Send email on first publish to test (featured-flagged)
https://trello.com/c/te7rYMDe

Now that we have the Notify integration in place, we can hook the sending of the email and the rules for this to trigger.

Given we really don't want this email to be sent if the service couldn't be published due to some configuration error, or k8s commands error, it will be triggered as an `after_perform` callback on the `PublishServiceJob`, ensuring it only gets called if the job has finalised successfully.

Additionally this new service object has been called `AfterCompletion` as there might be other code we want to run here as well, but for now is only the email.

Note it is still behind feature flag so will not trigger on live editor, only in test/testable editors.