Fix 'undo move' button height
https://trello.com/c/PYyU1n7o

The `fb-govuk-button-inverted` class applies top and bottom padding of 5px, when the `govuk-button` has 8px / 7px top/bottom padding.

Removed the box-shadow on the blue button, as otherwise it will make the bottom border look too thick and will not align with the undo button.

Applied to the undo button the `govuk-button` class so it has all the base styles.