Skip acceptance tests Notify mail deliveries (#2648)
* Skip acceptance tests Notify mail deliveries

https://trello.com/c/OoqBKxm5

Although it is only a nuisance, every time the acceptance tests run there are some tests that will attempt to trigger Notify emails, but the API key for the test environment does not allow sending to the email recipient and returns an error that gets reported to Sentry.

This PR will intercept emails unless we are on `formbuilder-saas-live`. If the recipient is the acceptance tests user, will stop the delivery, meaning we will not call the Notify API so no error will be produced.

* Fallback when ENV is not present

On assets precompile (docker build) we don't have this ENV yet, as this env is injected into the pods by k8s.

Set a fallback.

* Refactor to simplify code

`ACCEPTANCE_TESTS_USER` is not exposed in the pods, where we need it.

At this point is easier to just use the hardcoded email address, and also refactor to make it simpler to add more email recipients to skip in the future.